winex11: Fix sub-pixel mouse raw motion
Accumulate sub-pixel (fraction of a pixel) mouse raw motion in a double,
and flush/send the integer part to Windows' INPUT when ready.

Those sub-pixel motion were lost because of the cast to
integer (XIRawEvent's values are doubles, INPUT's are LONG). It seems to
happen more with high dpi mice.

The original bug: in some games, very slow mouse motion seemed
unregistered. In The Witcher 3, with a 6000 dpi mouse: the mouse would
randomly hit invisible walls, but only sometimes, and when moving
somewhat slowly. Often the bug looked like mouse stutter or random
short mouse freeze.