ntdll: Refactor LdrInitializeThunk.
Since LdrInitializeThunk is exported by ntdll, its semantics may be relied on
by programs. At least one program expects that it is called whenever a new
thread is created and that the first argument is the CONTEXT for the thread.
Furthermore, it is expected that on i386 Eax will contain the thread's entry
point.